[image] apple needs explicit cast

+++ src/aliceVision/image/io.cpp
@@ -573,7 +573,7 @@ void readImage(const std::string& path,
                 ALICEVISION_THROW_ERROR("Working color space must be set to \"no_conversion\" if raw color interpretation is set to \"none\"");
             }
 
-            float user_mul[4] = {1,1,1,1};
+            float user_mul[4] = {1.f, 1.f, 1.f, 1.f};
 
             configSpec.attribute("raw:auto_bright", 0); // disable exposure correction
             configSpec.attribute("raw:use_camera_wb", 0); // no white balance correction
@@ -615,7 +615,8 @@ void readImage(const std::string& path,
             {
                 ALICEVISION_THROW_ERROR("A DCP color profile is required but cannot be found");
             }
-            float user_mul[4] = { neutral[0],neutral[1],neutral[2],neutral[1] };
+            float user_mul[4] = {static_cast<float>(neutral[0]), static_cast<float>(neutral[1]),
+                                 static_cast<float>(neutral[2]), static_cast<float>(neutral[1])};
             if (imageReadOptions.doWBAfterDemosaicing)
             {
                 for (int i = 0; i < 4; ++i)
@@ -640,7 +641,8 @@ void readImage(const std::string& path,
             {
                 ALICEVISION_THROW_ERROR("A DCP color profile is required but cannot be found");
             }
-            float user_mul[4] = { neutral[0],neutral[1],neutral[2],neutral[1] };
+            float user_mul[4] = {static_cast<float>(neutral[0]), static_cast<float>(neutral[1]),
+                                 static_cast<float>(neutral[2]), static_cast<float>(neutral[1])};
             if (imageReadOptions.doWBAfterDemosaicing)
             {
                 for (int i = 0; i < 4; ++i)
